ci: Set Up Python Testing Workflow (#36)
# Description

This change introduces a new GitHub Actions workflow for code testing and enhances the project's Justfile with additional commands. The new workflow, `code-test.yml`, sets up an environment to run unit tests on push to main and pull request events. It uses Poetry for dependency management and Just for running commands.

fixes #45