ci: Disable Natural Language Validation in Code Quality Workflow (#63)
# Description

This change disables the natural language validation in the code quality
GitHub Actions workflow. Specifically, it adds the
`VALIDATE_NATURAL_LANGUAGE: false` flag to the existing list of disabled
validations in the `super-linter` step. This modification ensures that
natural language checks are not performed during the code quality
assessment process.

fixes #65