ci: Update coverage path in unit test command (#77)
# Description

This change modifies the unit test command in the Justfile to expand the
coverage scope. Instead of limiting coverage to the `checker` directory,
it now covers the entire project root (`.`). This adjustment ensures
that the test coverage report includes all relevant code within the
project, providing a more comprehensive overview of the test coverage.

fixes #78