build: Enforce Action Inputs (#170)
# Pull Request

## Description

This pull request includes several changes to improve the configuration
handling and testing in the `checker` module. The most important changes
include updating the `Configuration` class to ensure configuration
values are checked, modifying the `docker-run` command to use the
configuration file path, and adding new tests for the `Configuration`
class.

Improvements to configuration handling:

*
[`checker/configuration.py`](diffhunk://#diff-34fa747a1de0f0db4a41ccbad0672567a39e82d3391ecd4de0d47398fc940a9bL11-R28):
Updated the `Configuration` class to include a new method
`get_and_check_for_value` for retrieving and validating configuration
values. This method raises a `ValueError` if the configuration value is
not set.

Changes to Docker commands:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL47-R48):
Modified the `docker-run` command to include an environment variable
`INPUT_CONFIG_FILE_PATH` with the path to the configuration file.

Enhancements to testing:

*
[`checker/tests/test_checker.py`](diffhunk://#diff-fc185135883e7db9de3ebdab2c44b8af927518dc2765e19221f225bba97a8c2eR8-R16):
Added a mock for the `Configuration` class in the `test_run_checker`
function to ensure it is called during the test.
*
[`checker/tests/test_configuration.py`](diffhunk://#diff-abd3d2422a5089670dbfb20640bc8f64b7a0e9aa8a5534c799094f872f67096bR1-R13):
Added a new test `test_configuration` to verify that the `Configuration`
class correctly retrieves and validates the configuration file path from
the environment variables.

Type hints and imports:

*
[`checker/configuration.py`](diffhunk://#diff-34fa747a1de0f0db4a41ccbad0672567a39e82d3391ecd4de0d47398fc940a9bR2):
Added the `Self` type hint from the `typing` module to improve type
annotations in the `get_and_check_for_value` method.

Fixes #111