build: Add Output File Path (#190)
# Pull Request

## Description

This pull request includes updates to the documentation and
configuration of the GitHub Action for the project status checker. The
changes primarily focus on enhancing the README and `action.yml` files
to provide better clarity and additional configuration options.

Documentation updates:

* Added a new section for "Inputs" in the README to describe the
configuration options available for the GitHub Action.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R21)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R33-R44)

Configuration updates:

* Updated `action.yml` to include a new input `output_file_path` with a
description and default value. Also, improved the description for the
existing `config_file_path` input.

Fixes #189