ci: Add Zizmor Job (#235)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for running a security check with zizmor. The most important change is
the addition of the `run-zizmor` job to the
`.github/workflows/code-checks.yml` file.

### Additions to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R163-R187):
Added a new job `run-zizmor` to check GitHub Actions with zizmor,
including steps for checking out the repository, installing the latest
version of `uv`, running zizmor, and uploading the SARIF file.

Fixes #234