chore: Set some job runners to ubuntu-22.04 (#258)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflows to
ensure compatibility and stability by updating the runner version.

Updates to GitHub Actions workflows:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L82-R82):
Updated the `runs-on` parameter for the `check-markdown-links` job to
use `ubuntu-22.04` instead of `ubuntu-latest`.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L196-R196):
Updated the `runs-on` parameter for the `run-local-action` job to use
`ubuntu-22.04` instead of `ubuntu-latest`.