ci: Add Pyproject Checks (#259)
# Pull Request

## Description

This pull request introduces a new validation step for the
`pyproject.toml` file in the CI workflow and adds a corresponding
command in the `Justfile`.

Enhancements to CI workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R69-R70):
Added a step to validate the `pyproject.toml` file using the `just
pyproject-check` command.

New command in `Justfile`:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR23-R26):
Added a `pyproject-check` command that uses `poetry check` to validate
the `pyproject.toml` file.