ci: Update Dependency Review (#274)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflow
configuration to improve permissions and add a new feature to the
dependency review process.

Improvements to GitHub Actions workflow:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR49-R50):
Added `pull-requests: write` permission to the `dependency-review` job
to allow it to write comments on pull requests.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR59-R60):
Configured the `dependency-review` action to comment a summary in the
pull request only on failure.