chore: Update Labels (#284)
# Pull Request

## Description

This pull request includes updates to the labeling configuration for
pull requests and the associated labels. The changes add new label
categories and reorganize existing ones to improve the labeling system.

Updates to labeling configuration:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R2-R20):
Added new label categories for `dependencies`, `configurations`, and
`typescript`. Reorganized existing labels and updated the
`github_actions` and `just` labels.
[[1]](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R2-R20)
[[2]](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9L16-R56)

Updates to labels:

*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR5-R7):
Added new labels for `configurations` and `typescript`. Re-added and
updated the `just` and `python` labels.
[[1]](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR5-R7)
[[2]](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dL47-R68)