ci: Add Lefthook Validate Job (#289)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
validate Lefthook configurations. The most important change is the
addition of the `lefthook-validate` job in the
`.github/workflows/code-checks.yml` file.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R193-R208):
Added a new job `lefthook-validate` to validate Lefthook configurations.
This job checks out the repository, installs the latest version of `uv`,
and runs `lefthook validate`.