chore: Update Dependabot Schedule (#321)
# Pull Request

## Description

This pull request modifies the Dependabot configuration to standardize
the update schedules for various dependency groups by switching from a
daily interval to a cron-based schedule.

**Standardization of update schedules:**

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L12-R13):
Updated the `github-actions` group to use a cron-based schedule with the
expression `"30 7 * * *"` instead of the previous daily interval.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L29-R29):
Updated the `docker` group to use a cron-based schedule with the
expression `"30 7 * * *"` instead of the previous daily interval.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L46-R45):
Updated the `python` group to use a cron-based schedule with the
expression `"30 7 * * *"` instead of the previous daily interval.