chore: Set Python Version (#342)
# Pull Request

## Description

This pull request updates dependencies and adjusts compatibility
specifications in the `Dockerfile` and `pyproject.toml` to ensure the
project remains up-to-date and compatible with the required tools and
libraries.

### Dependency Updates:
* Updated `uv` to version 0.7.5 in both the `Dockerfile` and
`pyproject.toml` to align with the latest required version.
(`Dockerfile`
[[1]](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557L8-R8)
`pyproject.toml`
[[2]](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711L14-R22)
* Updated `ruff` to version 0.11.10 in the development dependencies of
`pyproject.toml`.

### Compatibility Adjustments:
* Changed the `requires-python` specification in `pyproject.toml` from
`>=3.13` to `~=3.13` to restrict compatibility to Python 3.13.x
versions.