ci: Update Pull Request Tasks Permissions (#436)
# Pull Request

## Description

This pull request makes a minor adjustment to the GitHub Actions
workflow configuration to move the `pull-requests: read` permission from
the global workflow level to the specific job that requires it.

- Workflow permissions update:
* In `.github/workflows/pull-request-tasks.yml`, the global
`permissions` block was changed to an empty object, and the
`pull-requests: read` permission was moved into the
`check-pull-request-title` job, restricting its scope to only that job.